### 2022-12-01 #4987 @editorialbot
[REVIEW]: signnet: An R package for analyzing signed networks
**Submitting author:** @schochastics (<a href="http://orcid.org/0000-0003-2952-4812">David Schoch</a>)
**Repository:** https://github.com/schochastics/signnet
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
**Archive:** https://doi.org/10.5281/zenodo.7522563
