### 2022-12-02 #4987 @zpneal
This is a much-needed package and very clear description of its features. I only have a few minor suggestions:

Installation - The package installs fine (I am reviewing the development version from GitHub). However, I suggest including the installation syntax directly in the manuscript. For example, maybe insert `install.packages("signnet")`  (or similar, if it must be installed from GitHub) right before `library(signnet)` on page 1.

Reproducibility & Functionality -
When running `frustration_exact(tribes)` I get the error `the package 'ompr' is needed for this function to work`. Currently the ompt package is only suggested, but not imported. However, if it is critical to run `frustration_exact()`, which is such an important function that it is demonstrated in this paper, then I suggest it be imported.

State of the field -
You explain that computing the optimal partition that minimizes frustration is NP hard and can only be calculated for small signed networks. While it is NP hard, and perhaps the `signnet` implementation only works for small signed networks, Aref & Neal (2020, 2021 in Scientific Reports) demonstrate a method applied to much larger signed networks. I suggest either removing this claim, or adding a reference to this work.

You cite the R `backbone` package when you discuss the fact that a weighted projection can be converted into an unweighted network. It may also be worth noting, either here or elsewhere, that the `backbone` package can generate signed backbones (from projections, but also from unipartite graphs) in an `igraph` format that is compatible with `signnet` for subsequent analysis. This may highlight to readers one possible source of signed network data, since they may be able to generate signed networks from existing non-signed networks.

References -
You currently cite Domagalski et al (2021), however the `backbone` package has an updated reference. I suggest replacing this with Neal (2022, https://doi.org/10.1371/journal.pone.0269137).
