### 2022-12-02 e631ffb
addressed review https://github.com/openjournals/joss-reviews/issues/4987#issuecomment-1335336761
