+++ paper.bib
@@ -215,3 +215,38 @@
   publisher = {{arXiv}},
   doi = {10.48550/arXiv.2205.13411}
 }
+
+@article{an-dcopsnpc-20,
+  title = {Detecting Coalitions by Optimally Partitioning Signed Networks of Political Collaboration},
+  author = {Aref, Samin and Neal, Zachary},
+  year = {2020},
+  journal = {Scientific Reports},
+  volume = {10},
+  number = {1},
+  pages = {1506},
+  publisher = {{Nature Publishing Group}},
+  issn = {2045-2322},
+  doi = {10.1038/s41598-020-58471-z}
+}
+
+@manual{s-omsmilp-22,
+  title = {ompr: Model and Solve Mixed Integer Linear Programs},
+  author = {Dirk Schumacher},
+  year = {2022},
+  note = {R package version 1.0.3},
+  url = {https://CRAN.R-project.org/package=ompr},
+}
+
+@article{n-brpenb-22,
+    doi = {10.1371/journal.pone.0269137},
+    author = {Neal, Zachary P.},
+    journal = {PLOS ONE},
+    publisher = {Public Library of Science},
+    title = {backbone: An {{R}} package to extract network backbones},
+    year = {2022},
+    month = {05},
+    volume = {17},
+    url = {https://doi.org/10.1371/journal.pone.0269137},
+    pages = {1-24},
+    number = {5}
+}
