### 2022-12-02 #4987 @schochastics
> install.packages("signnet") (or similar, if it must be installed from GitHub) right before library(signnet) on page 1.

added. Good that you bring this up: The CRAN version is outdated at the moment but after this review process I am going to submit the new version again. So the version to be reviewed is on Github but eventually it will be on CRAN so I added `install.packages("signnet")`

>Reproducibility & Functionality -
When running frustration_exact(tribes) I get the error the package 'ompr' is needed for this function to work. Currently the ompt package is only suggested, but not imported. However, if it is critical to run frustration_exact(), which is such an important function that it is demonstrated in this paper, then I suggest it be imported.

Happy to discuss this point, but I would like to keep it under Suggests. The ompr ecosystem includes several packages which I do not want to force onto every user. Of course the exact calculation is important, but I'd rather have those who want to use it go the extra mile given their might be quirks in the installation. I did add a sentence now explaining that more packages are needed to use the function. But as I said, I can still be convinced otherwise :)

>State of the field
You explain that computing the optimal partition that minimizes frustration is NP hard and can only be calculated for small signed networks. While it is NP hard, and perhaps the signnet implementation only works for small signed networks, Aref & Neal (2020, 2021 in Scientific Reports) demonstrate a method applied to much larger signed networks. I suggest either removing this claim, or adding a reference to this work.

I think I went on auto pilot here. "NP hard-> small networks". Apologies for that. I did implement the efficient algorithm but the solver is not as efficient as the one used bei Samin, I believe. I did add a reference and also changed the sentence.

>You cite the R backbone package when you discuss the fact that a weighted projection can be converted into an unweighted network. It may also be worth noting, either here or elsewhere, that the backbone package can generate signed backbones (from projections, but also from unipartite graphs) in an igraph format that is compatible with signnet for subsequent analysis. This may highlight to readers one possible source of signed network data, since they may be able to generate signed networks from existing non-signed networks.

good point! I added a sentence at the end of the Statement of need

> References
You currently cite Domagalski et al (2021), however the backbone package has an updated reference. I suggest replacing this with Neal (2022, https://doi.org/10.1371/journal.pone.0269137).

changed
