### 2022-12-02 #4987 @zpneal
Thanks for these revisions. (FYI - There's an extra quotation mark in the new line to install signnet)

Re: ompr - I agree about keeping required dependencies to a minimum. I think keeping this as suggests rather than imports makes sense, but this means the code in the paper won't replicate without some extra user steps. I suggest adding to the paper the extra `install.packages("ompr")` line so that the `frustration_exact(tribes)' function will replicate. You could add a note here explaining that `ompr` is only needed for certain package functions (which ones?), and so is not installed by default.

With that change, I can also tick off the "reproducibility" item from the checklist and would be satisfied with the submission.
