### 2022-12-02 653cd8c
made suggested changes https://github.com/openjournals/joss-reviews/issues/4987#issuecomment-1335462954
