### 2022-12-02 #4987 @schochastics
Thats of course a good point. I fixed the extra quotation and also now added the installation step.
