### 2022-12-02 #4987 @cosimameyer
That was a very fast review, Zachary :) I’ll do my best to finish mine also as fast as possible. Just a quick note, as mentioned in the other issue, I‘m currently on holiday and will only get a chance to look into the package at the end of next week - but I‘m looking forward to digging into it! It looks very cool! 😊
