### 2022-12-19 schochastics/signnet#11 @cosimameyer
[Suggestion] Add border around your visualization in README.md
I love your visualization and the explanation in your README.md. But since I am using dark mode in my browser, I find it hard to read your first visualization (see screenshot). This is only a suggestion, but if you had the time, it would be fantastic to add a border around the visualization to make it readable 😊

(part of https://github.com/openjournals/joss-reviews/issues/4987)
