### 2022-12-19 schochastics/signnet#11 @schochastics
Oh, thats bad because the text is also hidden. I added a white background now which should fix the readability issues.
fyi the figure is made with [roughnet](https://github.com/schochastics/roughnet).
