### 2022-12-21 schochastics/signnet#12 @cosimameyer
Add code snippets to reproduce triad_census_signed() visualization in README.md
For beginners (or users in general), I think adding code snippets to reproduce the `triad_census_signed()` in the README.md would be a great asset (similar to the `ggblock()` examples).

The code snippets are missing here:

(This issue is part of the review https://github.com/openjournals/joss-reviews/issues/4987)
