### 2022-12-21 schochastics/signnet#13 @cosimameyer
Add and extend contributing guidelines to README.md
Specific ontributor guidelines exist on the [package's website](http://signnet.schochastics.net/CONTRIBUTING.html). Making them also transparent in your GitHub README.md would be fantastic (I think most people will use the GitHub repository as the landing page to explore more information on the package and there's currently only a general "Code of Conduct" that does not address where to address how to explicitly "1) Contribute to the software 2) Report issues or problems with the software 3) Seek support").

Similarly, making clear how to seek support (for instance, by opening an issue) would provide (new) users with good guidance. It's a suggestion, but we added it to our package README [this way](https://github.com/cosimameyer/overviewR#how-to-reach-out) throughout the review process at JOSS and I think it was a good suggestion by our reviewer.

(This issue is part of the review https://github.com/openjournals/joss-reviews/issues/4987)
