### 2022-12-21 schochastics/signnet#14 @cosimameyer
[Documentation] One file per function (and documentation also for all helper functions)
It's probably a matter of taste but I like having one R file per function (where the R file has the same name as the function). This makes the code basis easily accessible for new users and also facilitates debugging (the following references are based on the development version of your package that is currently on GitHub):

- You have multiple functions in `signed_triangles.R`, `blockmodel.R`, `centrality_indices.R`, `complex_matrices.R`, `blockmodel.R`, `balance_scores.R`, `plot.R`, `random_graphs.R`

 When debugging your package, I also saw that some functions are not documented:

- You don't have documentation for `genclu()` (stored in `blockmodel.R`)
- (Undocumented) helper functions (`circleFun()` and `arc_dist()`; stored in `random_graphs.R` - but probably rather belong to `utils.R`? At least from a user perspective, it would make more sense to have them there if you don't want to have them in separate files)

(This issue is part of the review: https://github.com/openjournals/joss-reviews/issues/4987)
