### 2022-12-21 schochastics/signnet#15 @cosimameyer
Remove commented code blocks in functions
I saw some commented code blocks in the following R files:

- `balance_scores.R`
- `laplace_matrix.R` (and here the inactive `laplacian_angle()` function)
- `signed_triangles.R`

To keep the code clean, I'd remove the unused snippets (and make use of Git in case you want to use the functions/code in the future 😊)

(This issue is part of the review: https://github.com/openjournals/joss-reviews/issues/4987; the references above are based on the development version of your package that is currently on GitHub)
