### 2022-12-21 schochastics/signnet#16 @cosimameyer
Mismatch between package versions applied throughout the paper
When reproducing the results presented in the paper, I experienced a small hic-up with the package versions. In the paper, you install the CRAN version (0.8.1) but use `graph_from_adjacency_matrix_signed()` in the "Centrality" section (which was, based on [your release statements](http://signnet.schochastics.net/news/index.html), introduced in 1.0.0 and currently the development version). To replicate the results, you'd need to also introduce the development package version call here.

When using version 1.0.0, however, I'm running into (exporting) errors and cannot reproduce the following code snippet from your paper:

```
g <- graph_from_adjacency_matrix_signed(A, "undirected")
```

It looks like the function was not correctly exported. I'm not 100% sure why this happens. You export the function `graph_from_adjacency_matrix_signed ()` with `@export` and it's also listed in NAMESPACE  - so everything looks good from this perspective. Within the `signnet` R-project package environment, it also works perfectly after loading the package. But when calling the function with `signnet::` (or even `signnet:::` - just to eliminate all options) in another project, I get the following errors:

```
devtools::install_github("schochastics/signnet")
library(signnet)

g <- graph_from_adjacency_matrix_signed(A, "undirected")
# Error in graph_from_adjacency_matrix_signed(A, "undirected") :
#  could not find function "graph_from_adjacency_matrix_signed"

g <- signnet::graph_from_adjacency_matrix_signed(A, "undirected")
# Error: 'graph_from_adjacency_matrix_signed' is not an exported object from 'namespace:signnet'
```

Calling `ls("package:signnet")` returns the following output where `graph_from_adjacency_matrix_signed ` is not listed:

```
 [1] "as_adj_complex"            "as_adj_signed"             "as_complex_edges"
 [4] "as_incidence_complex"      "as_incidence_signed"       "as_signed_proj"
 [7] "as_unsigned_2mode"         "avatar"                    "balance_score"
[10] "complex_walks"             "count_complex_triangles"   "count_signed_triangles"
[13] "cowList"                   "degree_signed"             "eigen_centrality_signed"
[16] "frustration_exact"         "ggblock"                   "ggsigned"
[19] "graph_circular_signed"     "laplacian_angle"           "laplacian_matrix_complex"
[22] "laplacian_matrix_signed"   "pn_index"                  "sample_islands_signed"
[25] "signed_blockmodel"         "signed_blockmodel_general" "signed_triangles"
[28] "triad_census_signed"       "tribes"
```

Here is my session info:

```
─ Session info  setting  value
 version  R version 4.1.3 (2022-03-10)
 os       macOS 13.0.1
 system   x86_64, darwin17.0
 ui       RStudio
 language (EN)
 collate  en_US.UTF-8
 ctype    en_US.UTF-8
 tz       Europe/Berlin
 date     2022-12-21
```

(This issue is part of the review: https://github.com/openjournals/joss-reviews/issues/4987)
