### 2022-12-21 schochastics/signnet#17 @cosimameyer
[Suggestion] Increase code coverage in unit tests for remaining functions
In general, the package contains automated tests and your code coverage is high -- I appreciate it! 🙂

This is only a suggestion but I believe that a well-documented code also comes with good code coverage. When checking your coverage, I saw that you that there are a few functions that have little or no code coverage and it would be great, if you can also implement unit tests for these functions:

(This issue is part of the review: https://github.com/openjournals/joss-reviews/issues/4987; the references above are based on the development version of your package that is currently on GitHub)
