### 2022-12-21 schochastics/signnet#18 @cosimameyer
[Suggestion] Quality checks/Code Hygiene
When developing R packages, I usually run a couple of functions that are helpful to automatically check different dimensions of quality (spelling, linter, ...). I used the following functions and they threw a couple of warnings/errors you might want to have a look at:

[Source for screenshot with full list of checks](https://www.mzes.uni-mannheim.de/socialsciencedatalab/article/r-package/#section6)

The warnings included:

- Spelling errors
- Increase unit tests for code coverage
- Use `TRUE`/`FALSE` instead of `T`/`F`
- Avoid `1:length()` (error prone)
- `fix this R CMD check ERROR: Package suggested but
    not available: ‘ggraph’ The suggested packages are required
    for a complete check.`
- Typical linting errors (add spaces, avoid long code lines)
- Your DESCRIPTION does not adhere to the "tidy description" standards (if you want to follow the conventions, run `usethis::use_tidy_description()`)

(This issue is part of the review: https://github.com/openjournals/joss-reviews/issues/4987; the references above are based on the development version of your package that is currently on GitHub)
