### 2022-12-21 schochastics/signnet#19 @cosimameyer
[Suggestion] Set package name in paper always in code block
When going through the paper, I recognized a mismatch in the fonts that were used to refer to `igraph`. Within one paragraph, the package was addressed as `igraph` and igraph (see screenshot) - and I'm not sure why. For consistency, I recommend setting package references in code blocks and referring to the package as `igraph`.

(This issue is part of the review: https://github.com/openjournals/joss-reviews/issues/4987)
