### 2022-12-21 #4987 @cosimameyer
Thank you very much for the opportunity to review the software and the package!

I added a few issues (also including suggestions). I enjoyed going through the software and the paper more in-depth and hope my comments help to make an even better version of the package.

Although not included in the paper, I'd like to highlight the hands-on tutorials on the [package website](http://signnet.schochastics.net/index.html) and would like to encourage David to also refer the interested reader to them within the paper 🙂
