### 2022-12-21 schochastics/signnet#19 @schochastics
Seems I forgot to put igraph between backticks on several occasions. Should be fixed now
