### 2022-12-21 c5306e4
added backticks around igraph (#19)
