### 2022-12-21 fed6d6f
Quality checks/code hygiene (#18)
