+++ man/as_adj_complex.Rd
@@ -9,7 +9,7 @@ as_adj_complex(g, attr)
 \arguments{
 \item{g}{igraph object}
 
-\item{attr}{edge attribute name that encodes positve ("P"), negative ("N") and ambivalent ("A") ties.}
+\item{attr}{edge attribute name that encodes positive ("P"), negative ("N") and ambivalent ("A") ties.}
 }
 \value{
 complex adjacency matrix
