### 2022-12-21 schochastics/signnet#18 @schochastics
fixed what I was able fix, except the code coverage (see review):
- spell check now only returns words that it is not familiar with
- All `T`/`F` where exchanged with `TRUE`/`FALSE`
- removed instances `1:length`
