### 2022-12-21 schochastics/signnet#17 @schochastics
see discussion in the review
