### 2022-12-21 schochastics/signnet#16 @schochastics
see discussion in review
