### 2022-12-21 a15fdd6
removed commented code (fix #15)
