+++ R/signed_triangles.R
@@ -434,175 +434,3 @@ triad_census_signed <- function(g) {
   census[1] <- choose(n, 3) - sum(census)
   census
 }
-
-
-
-# orbits <- c("35-35-35", "34-32-33", "35-35-35", "34-33-32", "25-26-26",
-#             "34-33-32", "35-35-35", "34-32-33", "35-35-35", "32-34-33", "19-19-20",
-#             "32-34-33", "29-31-30", "13-15-14", "29-31-30", "32-34-33", "19-19-20",
-#             "32-34-33", "35-35-35", "34-32-33", "35-35-35", "34-33-32", "25-26-26",
-#             "34-33-32", "35-35-35", "34-32-33", "35-35-35", "33-34-32", "31-29-30",
-#             "33-34-32", "28-28-27", "22-24-23", "28-28-27", "33-34-32", "31-29-30",
-#             "33-34-32", "26-25-26", "15-13-14", "26-25-26", "24-22-23", "12-12-11",
-#             "24-22-23", "26-25-26", "15-13-14", "26-25-26", "33-34-32", "31-29-30",
-#             "33-34-32", "28-28-27", "22-24-23", "28-28-27", "33-34-32", "31-29-30",
-#             "33-34-32", "35-35-35", "34-32-33", "35-35-35", "34-33-32", "25-26-26",
-#             "34-33-32", "35-35-35", "34-32-33", "35-35-35", "32-34-33", "19-19-20",
-#             "32-34-33", "29-31-30", "13-15-14", "29-31-30", "32-34-33", "19-19-20",
-#             "32-34-33", "35-35-35", "34-32-33", "35-35-35", "34-33-32", "25-26-26",
-#             "34-33-32", "35-35-35", "34-32-33", "35-35-35", "32-33-34", "29-30-31",
-#             "32-33-34", "19-20-19", "13-14-15", "19-20-19", "32-33-34", "29-30-31",
-#             "32-33-34", "27-28-28", "16-18-17", "27-28-28", "16-17-18", "4-5-5",
-#             "16-17-18", "27-28-28", "16-18-17", "27-28-28", "32-33-34", "29-30-31",
-#             "32-33-34", "19-20-19", "13-14-15", "19-20-19", "32-33-34", "29-30-31",
-#             "32-33-34", "30-31-29", "21-21-21", "30-31-29", "18-17-16", "8-6-7",
-#             "18-17-16", "30-31-29", "21-21-21", "30-31-29", "23-22-24", "7-8-6",
-#             "23-22-24", "10-9-10", "1-3-2", "10-9-10", "23-22-24", "7-8-6",
-#             "23-22-24", "30-31-29", "21-21-21", "30-31-29", "18-17-16", "8-6-7",
-#             "18-17-16", "30-31-29", "21-21-21", "30-31-29", "32-33-34", "29-30-31",
-#             "32-33-34", "19-20-19", "13-14-15", "19-20-19", "32-33-34", "29-30-31",
-#             "32-33-34", "27-28-28", "16-18-17", "27-28-28", "16-17-18", "4-5-5",
-#             "16-17-18", "27-28-28", "16-18-17", "27-28-28", "32-33-34", "29-30-31",
-#             "32-33-34", "19-20-19", "13-14-15", "19-20-19", "32-33-34", "29-30-31",
-#             "32-33-34", "35-35-35", "34-32-33", "35-35-35", "34-33-32", "25-26-26",
[…]
