### 2022-12-21 6125140
added description to undocumented functions (#14)
