### 2022-12-21 schochastics/signnet#14 @schochastics
Indeed I think this is a matter of taste. I personally like to keep functions of the same category within one file.
There does not seem to be an agreed standard, but this is what the [R packages](https://r-pkgs.org/) book says:
![rpkg](https://user-images.githubusercontent.com/17147355/208995027-38d790b0-9e84-453a-bd86-5f837bf8f786.png)

I wrote a short description for the undocumented functions but since they are not exported, I kept it on the brief side.
