### 2022-12-21 aa54856
added contribution guide (fixes #13)
