+++ README.md
@@ -138,7 +138,7 @@ blockmat
 clu <- signed_blockmodel_general(g,blockmat,0.5)
 clu
 #> $membership
-#>  [1] 2 2 2 3 2 1 1 1 1 1 1 3 3 3 1
+#>  [1] 1 1 1 1 1 2 2 2 2 2 3 3 3 3 3
 #> 
 #> $criterion
 #> [1] 0
@@ -147,9 +147,29 @@ ggblock(g,clu$membership,show_blocks = TRUE,show_labels = FALSE)
 
 <img src="man/figures/README-general_example-1.png" width="100%" />
 
-## Code of Conduct
+# How to reach out?
 
-Please note that the netrankr project is released with a [Contributor
+### Where do I report bugs?
+
+Simply [open an
+issue](https://github.com/schochastics/signnet/issues/new) on GitHub.
+
+### How do I contribute to the package?
+
+If you have an idea (but no code yet), [open an
+issue](https://github.com/schochastics/signnet/issues/new) on GitHub. If
+you want to contribute with a specific feature and have the code ready,
+fork the repository, add your code, and create a pull request.
+
+### Do you need support?
+
+The easiest way is to [open an
+issue](https://github.com/schochastics/signnet/issues/new) - this way,
+your question is also visible to others who may face similar problems.
+
+### Code of Conduct
+
+Please note that the signnet project is released with a [Contributor
 Code of
 Conduct](https://contributor-covenant.org/version/2/1/CODE_OF_CONDUCT.html).
 By contributing to this project, you agree to abide by its terms.
