### 2022-12-21 7ad4aa3
better description for triad figure (#12)
