+++ README.Rmd
@@ -58,6 +58,7 @@ balance_score(tribes,method = "frustration")
 ```
 
 For directed signed networks, `triad_census_signed()` can be used to compute the count for all 138 non-isomorphic signed triads.
+*(The figure below shows all 138 signed triads that can occur in a directed network)*
 <img src="man/figures/signed_triads.png"/>
 
 # Blockmodeling
