### 2022-12-21 schochastics/signnet#12 @schochastics
The code for the `ggblock` figures is already in the README. I see that there is a bit of a confusion for the `triad_census_signed()`. That function produces the count of each of the 138 triads than can occur. The figure is just there to shows what these triads look like and cannot be reproduced with the package. I added a short description now which hopefully makes this clear.
