### 2022-12-21 #4987 @schochastics
@cosimameyer  Thanks for thoroughly digging into the package!
I was able to close some of the issues but some might need your approval/more discussion.
