### 2022-12-22 schochastics/signnet#12 @cosimameyer
Sorry for the confusion here that may be caused by comment.
I was actually looking for a code snippet to reproduce the figure similar to the one you have already provided for `ggblock()`  🙂
For me it doesn't matter that much which package produces the figure that I added above - but as a user who goes through the README, I'd want to have code snippets that explain how you generated the figure (in case I want to replicate it). I hope this makes my suggestion clear 🙂
