### 2022-12-22 schochastics/signnet#17 @cosimameyer
Thanks, this absolutely works for me 🙂
