### 2022-12-22 schochastics/signnet#17 @cosimameyer
An idea to ease the testing would be to make use of [`xpectr`](https://github.com/LudvigOlsen/xpectr#functions-for-use-in-tests) (you probably already know about it - and it isn't meant to take away the active thinking for qualitative tests. But it least can generate tests that ensure the functional performance of your functions adheres to a specific type etc.).
Again, it won't take away more substantive and qualitative tests but it helps to generate systematic tests and works like this:

```
data(tribes)

xpectr::gxs_function(fn = frustration_exact,
             args_values = list(
               "g" = list(tribes)
             ))
```

which produces this (I added the data call):

```

## Testing 'frustration_exact'                                              ####
## Initially generated by xpectr
# Testing different combinations of argument values

# Testing frustration_exact(g = tribes)
xpectr::set_test_seed(42)

data(tribes)

# Assigning output
output_16692 <- frustration_exact(g = tribes)
# Testing class
expect_equal(
  class(output_16692),
  "list",
  fixed = TRUE)

# Testing type
expect_type(
  output_16692,
  type = "list")

# Testing values
expect_equal(
  output_16692[["frustration"]],
  7,
  tolerance = 1e-4)
expect_equal(
  output_16692[["partition"]],
  c(0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 0),
  tolerance = 1e-4)

# Testing names
expect_equal(
  names(output_16692),
  c("frustration", "partition"),
  fixed = TRUE)

# Testing length
expect_equal(
  length(output_16692),
  2L)

# Testing sum of element lengths
expect_equal(
  sum(xpectr::element_lengths(output_16692)),
  17L)

# Testing element classes
expect_equal(
  xpectr::element_classes(output_16692),
  c("numeric", "numeric"),
  fixed = TRUE)

# Testing element types
expect_equal(
  xpectr::element_types(output_16692),
  c("double", "double"),
  fixed = TRUE)

# Testing frustration_exact(g = NULL)
# Changed from baseline: g = NULL
xpectr::set_test_seed(42)
# Testing side effects
# Assigning side effects
side_effects_15156 <- xpectr::capture_side_effects(frustration_exact(g = NULL), reset_seed = TRUE)
expect_equal(
  xpectr::strip(side_effects_15156[['error']]),
  xpectr::strip("Not a graph object"),
  fixed = TRUE)
expect_equal(
  xpectr::strip(side_effects_15156[['error_class']]),
  xpectr::strip(c("simpleError", "error", "condition")),
  fixed = TRUE)

## Finished testing 'frustration_exact'                                     ####

```

(you'll need to go through it in detail and only take those tests that make sense but can be easily implemented in your test environment. I like `xpectr` a lot because it also helps you to capture possible type problems, for instance, that you didn't think explicitly about when writing your code 🙂)

You don't have to implement it in this run (it's only a suggestion) but probably it's helpful 🙂
