### 2022-12-22 schochastics/signnet#14 @cosimameyer
Thanks for following up - I know there is some controversy about it and I'm fine with leaving it as it is (again, as I said, it's a matter of taste).
Coming from a user experience perspective when you often work with new packages on a debugging modus, having the functions sorted in separate files (except for the helper functions that live in something like `utils.R`) really helps a lot but if you prefer to keep it as it is, it's also good for me 🙂
