### 2022-12-22 schochastics/signnet#16 @cosimameyer
I cannot fully explain why but I've got good news: it works now with v1.0.0 and I can reproduce your paper 🎉
(to better understand what happened: I re-ran the short script that I used yesterday - in a new R project today; I didn't use `packageVersion()` but the RStudio GUI to check for the package version (probably that was the problem that it was a bit laggy in RStudio's interface and didn't display the correct version number))

In order to get the paper accepted, you would need to slightly change the order of your plans: First submit v1.0.0 to CRAN (because this is the basis of your paper and you give instructions for an installation as if the version was already on CRAN) and then get the paper accepted 🙂

Otherwise you have a paper out there that cannot be replicated in its current version until CRAN approves the new version of your package.
