### 2022-12-22 schochastics/signnet#16 @schochastics
Yeah this makes far more sense than my original plans. I will submit it  to CRAN today and let you know when it is accepted there
