### 2022-12-22 schochastics/signnet#17 @schochastics
Thanks for the suggestion. Actually never worked with xpectr but i will take this opportunity to dive into it :slightly_smiling_face:
