### 2022-12-22 schochastics/signnet#16 @schochastics
@cosimameyer submitted and it was already accepted :open_mouth:
