### 2022-12-22 schochastics/signnet#16 @cosimameyer
That's fantastic news! (A bit like a Christmas miracle) 🥳 I'll check the tests tomorrow and if they all pass, I'm very happy to close this issue ☺️
