### 2022-12-22 d98ed6c
added gist for signed triads (#12)
