### 2022-12-22 schochastics/signnet#12 @schochastics
Ok now I got it :slightly_smiling_face: .
Since the code is a bit more complicated, I put it in a gist that is linked in the README.
