### 2023-01-10 #4987 @sbenthall
@schochastics Can you please:

- make a tagged release
- make an archive (on Zenodo for example), making sure that the archive has the correct metadata (title and author list)
-  and report the version number and archive DOI in this thread.
