### 2023-01-10 #4987 @schochastics
@sbenthall :

* tagged release: `v1.0.0` [link](https://github.com/schochastics/signnet/releases/tag/v1.0.0)
*  Zenodo archive: https://zenodo.org/record/7522563
* DOI: https://doi.org/10.5281/zenodo.7522563
