+++ README.md
@@ -11,6 +11,7 @@ status](https://www.r-pkg.org/badges/version/signnet)](https://cran.r-project.or
 [![Downloads](https://cranlogs.r-pkg.org/badges/signnet)](https://CRAN.R-project.org/package=signnet)
 [![Codecov test
 coverage](https://codecov.io/gh/schochastics/signnet/branch/master/graph/badge.svg)](https://app.codecov.io/gh/schochastics/signnet?branch=master)
+[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.7522563.svg)](https://doi.org/10.5281/zenodo.7522563)
 <!-- badges: end -->
 
 The package provides methods to analyse signed networks (i.e. networks
@@ -84,7 +85,7 @@ are negative.
 clu <- signed_blockmodel(tribes,k = 3,alpha = 0.5,annealing = TRUE)
 clu
 #> $membership
-#>  [1] 3 3 2 2 1 2 2 2 1 1 2 2 1 1 3 3
+#>  [1] 1 1 2 2 3 2 2 2 3 3 2 2 3 3 1 1
 #> 
 #> $criterion
 #> [1] 2
@@ -140,7 +141,7 @@ blockmat
 clu <- signed_blockmodel_general(g,blockmat,0.5)
 clu
 #> $membership
-#>  [1] 1 1 1 1 1 2 2 2 2 2 3 3 3 3 3
+#>  [1] 3 2 2 3 2 1 1 1 1 1 3 3 1 3 1
 #> 
 #> $criterion
 #> [1] 0
