+++ README.Rmd
@@ -19,6 +19,7 @@ knitr::opts_chunk$set(
 [![CRAN status](https://www.r-pkg.org/badges/version/signnet)](https://cran.r-project.org/package=signnet)
 [![Downloads](https://cranlogs.r-pkg.org/badges/signnet)](https://CRAN.R-project.org/package=signnet)
 [![Codecov test coverage](https://codecov.io/gh/schochastics/signnet/branch/master/graph/badge.svg)]( https://app.codecov.io/gh/schochastics/signnet?branch=master)
+[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.7522563.svg)](https://doi.org/10.5281/zenodo.7522563)
 <!-- badges: end -->
 
 The package provides methods to analyse signed networks (i.e. networks with both positive and negative ties).
