### 2023-01-10 #4987 @editorialbot
:wave: @openjournals/sbcs-eics, this paper is ready to be accepted and published.
