### 2023-01-16 8ad0ea7
fixed an error that occurs with new version of igraph (https://github.com/igraph/rigraph/pull/633)
