+++ R/signed_triangles.R
@@ -178,7 +178,7 @@ count_complex_triangles <- function(g, attr) {
 #' @examples
 #' library(igraph)
 #' g <- graph.full(4, directed = TRUE)
-#' E(g)$sign <- c(-1, 1, 1, -1, -1, 1)
+#' E(g)$sign <- rep(c(-1, 1, 1, -1, -1, 1),2)
 #' triad_census_signed(g)
 #' @export
 
