+++ man/triad_census_signed.Rd
@@ -18,7 +18,7 @@ triad census for signed graphs
 \examples{
 library(igraph)
 g <- graph.full(4, directed = TRUE)
-E(g)$sign <- c(-1, 1, 1, -1, -1, 1)
+E(g)$sign <- rep(c(-1, 1, 1, -1, -1, 1),2)
 triad_census_signed(g)
 }
 \author{
